Climate_controls_system/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Django==3.0.6
django-bootstrap4==1.1.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
paho-mqtt==1.5.0
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
vk-api==11.8.0
